chore: fix performance tests and improve speed
